# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: College Fest Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      parameters:
        - name: college_id
          in: query
          required: false
          schema:
            type: integer
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [upcoming, completed]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"

    post:
      operationId: createEvent
      tags: [events]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /superadmin/admins/{adminId}/approve:
    put:
      operationId: approveAdmin
      tags: [admin]
      parameters:
        - name: adminId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /superadmin/admins/{adminId}/reject:
    put:
      operationId: rejectAdmin
      tags: [admin]
      parameters:
        - name: adminId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /superadmin/admins/{adminId}/suspend:
    put:
      operationId: suspendAdmin
      tags: [admin]
      parameters:
        - name: adminId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        venue:
          type: string
        eventDate:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        maxParticipants:
          type: integer
          nullable: true
        registeredCount:
          type: integer
        thumbnailUrl:
          type: string
          nullable: true
        parentEventId:
          type: integer
          nullable: true
        previousEventId:
          type: integer
          nullable: true
        eventStatus:
          type: string
          enum: [upcoming, completed]
        collegeId:
          type: integer
        collegeName:
          type: string
        averageRating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, category, venue, eventDate, registrationDeadline, collegeId, collegeName, registeredCount, eventStatus, reviewCount, createdAt]
...
    CreateEventRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        venue:
          type: string
        eventDate:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        maxParticipants:
          type: integer
          nullable: true
        thumbnail:
          type: string
          format: binary
          nullable: true
        parentEventId:
          type: integer
          nullable: true
        previousEventId:
          type: integer
          nullable: true
      required: [title, description, category, venue, eventDate, registrationDeadline]
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        collegeName:
          type: string
        contactNumber:
          type: string
        designation:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending_verification, active, pending_approval, rejected, suspended]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, collegeName, contactNumber, status, createdAt]
...