# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * College Fest Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUserStatus } from "./adminUserStatus";

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  collegeName: string;
  contactNumber: string;
  designation?: string | null;
  status: AdminUserStatus;
  createdAt: Date;
}